+++ CHANGELOG.md
@@ -6,14 +6,16 @@ This file contains the list of changes made to the embc library.
 
 ## 0.2.1
 
-22 Oct 2020  [in progress]
+2020 Dec 15  [in progress]
 
 *   Fixed EMBC_TIME_TO_F32.
 *   Added embc_list_replace.
+*   Added stream data link and new framer.
+*   Improved unit tests and documentation.
 
 
 ## 0.1.1
 
-4 Feb 2017
+2017 Feb 4
 
 *   Initial public release.